make gShopAdvFlagMap a u16 to save space

+++ include/ItemIdMaps.hpp
@@ -1,8 +1,10 @@
 #pragma once
 
+#include <types.h>
+
 // 0x1F is the BMG group of rando.bmg
 #define MSG(idx) ((0x1F << 16) | (idx))
 
 extern u32 gGetItemMap[];
 extern u32 gBMGMap[];
-extern u32 gShopAdvFlagMap[];
+extern u16 gShopAdvFlagMap[];
